Testing.

+++ lab2/4.1/L1Cache.h
@@ -23,7 +23,7 @@ typedef struct CacheLine {
   uint8_t Valid;
   uint8_t Dirty;
   uint32_t Tag;
-  uint32_t Data[WORD_SIZE];
+  uint8_t Data[WORD_SIZE];
 } CacheLine;
 
 typedef struct Cache {
